requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
scikit-learn
requests
pydantic